refactor: split server routes/auth.js into modular route and lib files
Extracted shared helpers into server/lib/ and split the single route file
into three focused route files so each file has one responsibility.
